Rollup of Java fixes to add string/uint64 support & make the code saner.
1. Fix generated Java integer problem.

In the golden files, there are a number of properties with a type of "integer" but no format. In java_generator, this wasn't detected as a java.lang.Integer, but instead was treated like a class reference and sent through the class name cleanup, resulting in a class name of "Integer". Because Integer is in java.lang, it doesn't need to be imported or fully qualified. Things worked, though without fully qualified names.

And so this sat. Silent. Waiting. Story passed into legend and legend passed into myth and some things that never should have been forgotten passed out of the minds of mortals. Until one day, a young h̶o̶b̶b̶i̶t̶ engineer stumbled across it and it was super annoying that he had to handle both "java.lang.Integer" and "Integer", so he fixed it.  The end.

There's a ton of churn here, but it should all be "Integer" to "java.lang.Integer"

2. Remove 'Java12', leaving Java14 as equivalent to Java. The capability to eventually have a JavaX (X > 14) still exists.

- Merge Java14LanguageModel type defs up to JavaLanguageModel to combine strong
  import definitions and primitive types in a single type definition.
- Hoist j14lm type lookup methods up to JavaLangugeModel
- Remove unneeded method overrides form Java14LanguageModel.
- Remove java1_* lookups from generator_lookup.
- drive by cleanup: delete many obsolete java versions from the server targets.json.

Followup work:
- Change |('number',None)| from |Number| to |java.lang.Number|
- Raise an error on unknown type.

3. Start exposing batchPath from discovery.