Trim CRLF to LF in golden test output so that C# output is uniform and not dependent on who edited the .XML files which go into it.
Some files were created with CRLF in them, but the golden output for
those had been stripped to LF. Others not. Then depending on what
diff you use, you either see a change or don't. That's too confusing.

Note that the edits to the golden files are only of whitespace.