Add capability for better alignment with gapi for Javascript.
Added a reparent_methods_using_id flag to fix the case where methods with
IDs like api.method have paths like api.v1.method.  The gapi logic uses
the method ID, not the path, to determine the function name.  Since that
code is already used in running applications, we have to update the
externs generator to match it.